Refactoring: Getting rid of consul node manager - instead introducing consul node set.
